% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incase-package.R
\docType{package}
\name{incase-package}
\alias{incase}
\alias{incase-package}
\title{incase: Pipe-Friendly Vector Replacement with Case Statements}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Offers a pipe-friendly alternative to the 'dplyr'
    functions 'case_when()' and 'if_else()'.  These functions accept a
    vector as an optional first argument, allowing conditional statements
    to be built using the 'magrittr' dot operator.  The functions also
    coerce all possible outputs to the same type, meaning you no longer
    have to worry about using specific typed variants of NA or explicitly
    declaring integer outputs.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rossellhayes/incase}
  \item Report bugs at \url{https://github.com/rossellhayes/incase/issues}
}

}
\keyword{internal}
