% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incase-package.R
\docType{package}
\name{incase-package}
\alias{incase}
\alias{incase-package}
\title{incase: Pipe-Friendly Vector Replacement with Case Statements}
\description{
Offers a pipe-friendly alternative to \code{\link[dplyr:case_when]{dplyr::case_when()}} and
\code{\link[dplyr:if_else]{dplyr::if_else()}}.
These functions accept a vector as an optional first argument, allowing
conditional statements to be built using the \link[magrittr:magrittr-package]{magrittr}
dot operator.
The functions also coerce all possible outputs to the same type, meaning you
no longer have to worry about using specific typed variants of \code{NA} or
explicitly declaring integer outputs.
}
\section{Copyright}{
\strong{incase} includes code from \link[dplyr:dplyr-package]{dplyr}, which is
licensed under the MIT License:

Copyright (c) 2013-2019 RStudio and others.

Permission is hereby granted, free of charge, to any person obtaining a copy
of this software and associated documentation files (the "Software"), to deal
in the Software without restriction, including without limitation the rights
to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
copies of the Software, and to permit persons to whom the Software is
furnished to do so, subject to the following conditions:

The above copyright notice and this permission notice shall be included in
all copies or substantial portions of the Software.

THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
THE SOFTWARE.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://incase.rossellhayes.com}
  \item \url{https://github.com/rossellhayes/incase}
  \item Report bugs at \url{https://github.com/rossellhayes/incase/issues}
}

}
\author{
\strong{Maintainer}: Alexander Rossell Hayes \email{alexander@rossellhayes.com} (\href{https://orcid.org/0000-0001-9412-0457}{ORCID}) [copyright holder]

Other contributors:
\itemize{
  \item Patrice Kiener \email{fattailsr@inmodelia.com} (\href{https://orcid.org/0000-0002-0505-9920}{ORCID}) (Contributed example for fn_case()) [contributor]
}

}
\keyword{internal}
