% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switch_case.R
\name{switch_case}
\alias{switch_case}
\title{Switch-style recoding of values}
\usage{
switch_case(x, ..., preserve = FALSE, default = NA)
}
\arguments{
\item{x}{A vector}

\item{...}{<[`dynamic-dots`][rlang::dyn-dots]> A sequence of two-sided
  formulas.
  Elements of `x` that match the left hand side (LHS) will be replaced with
  the value in the right hand side (RHS).

  The LHS must evaluate to an atomic vector.

  The RHS must be of length one.

  `NULL` inputs are ignored.}

\item{preserve}{If `TRUE`, unmatched elements of `x` will be returned
unmodified.
(The elements may have their type coerced to be compatible with
replacement values.)
If `FALSE`, unmatched elements of `x` will be replaced with `default`.
Defaults to `FALSE`.}

\item{default}{If `preserve` is `FALSE`, a value to replace unmatched
elements of `x`.
Defaults to `NA`.}
}
\value{
A vector of the same length as `x`.
}
\description{
Switch-style recoding of values
}
\examples{
x <- 1:10
in_switch(
  x,
  3 ~ "fizz",
  5 ~ "buzz",
  6 ~ "fizz",
  9 ~ "fizz"
)

1:10 \%>\%
  in_switch(
    3 ~ "fizz",
    5 ~ "buzz",
    6 ~ "fizz",
    9 ~ "fizz"
  )

1:10 \%>\%
  in_switch(
    c(3, 6, 9) ~ "fizz",
    5 ~ "buzz"
  )

1:10 \%>\%
  in_switch(
    3 ~ "fizz",
    5 ~ "buzz",
    6 ~ "fizz",
    9 ~ "fizz",
    preserve = TRUE
  )

1:10 \%>\%
  in_switch(
    3 ~ "fizz",
    5 ~ "buzz",
    6 ~ "fizz",
    9 ~ "fizz",
    default = "bam"
  )
}
\seealso{
[in_case()], a pipeable alternative to [dplyr::case_when()]

  [if_case()], a pipeable alternative to [dplyr::if_else()]

  [switch()], which inspired this function
}
